requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1

Django~=5.0.4
pip~=24.0
attrs~=23.2.0
utils~=1.0.2
pillow~=10.3.0
automium_web~=0.1.1
asyncio~=3.4.3
Jinja2~=3.1.3
asgiref~=3.8.1
sqlparse~=0.4.4
psycopg2~=2.9.9
config~=0.5.1
colorama~=0.4.6
requests~=2.31.0
click~=8.1.7
typing_extensions~=4.11.0
views~=0.3
python-dotenv~=1.0.1
djangorestframework~=3.15.1